Refactor router implementation
